+++ arrow/request_test.go
@@ -129,7 +129,7 @@ func TestInsertRequestSetters(t *testing.T) {
 	buf := bytes.NewBuffer([]byte{})
 	enc := msgpack.NewEncoder(buf)
 
-	arr, err := arrow.MakeArrow([]byte{'a', 'b', 'c'})
+	arr, err := arrow.NewArrow([]byte{'a', 'b', 'c'})
 	require.NoError(t, err)
 
 	resolver := stubSchemeResolver{validSpace}