+++ crud/get.go
@@ -61,10 +61,11 @@ func (opts GetOpts) EncodeMsgpack(enc *msgpack.Encoder) error {
 // GetRequest helps you to create request object to call `crud.get`
 // for execution by a Connection.
 type GetRequest struct {
-	spaceRequest
+	baseRequest
 
-	key  Tuple
-	opts GetOpts
+	space string
+	key   Tuple
+	opts  GetOpts
 }
 
 type getArgs struct {
@@ -74,13 +75,13 @@ type getArgs struct {
 	Opts     GetOpts
 }
 
-// MakeGetRequest returns a new empty GetRequest.
-func MakeGetRequest(space string) GetRequest {
-	req := GetRequest{}
-	req.impl = newCall("crud.get")
-	req.space = space
-	req.opts = GetOpts{}
-	return req
+// NewGetRequest returns a new empty GetRequest.
+func NewGetRequest(space string) GetRequest {
+	return GetRequest{
+		baseRequest: newBaseRequest("crud.get"),
+		space:       space,
+		opts:        GetOpts{},
+	}
 }
 
 // Key sets the key for the GetRequest request.