+++ crud/insert_many.go
@@ -14,8 +14,9 @@ type InsertManyOpts = OperationManyOpts
 // InsertManyRequest helps you to create request object to call
 // `crud.insert_many` for execution by a Connection.
 type InsertManyRequest struct {
-	spaceRequest
+	baseRequest
 
+	space  string
 	tuples Tuples
 	opts   InsertManyOpts
 }
@@ -27,13 +28,13 @@ type insertManyArgs struct {
 	Opts     InsertManyOpts
 }
 
-// MakeInsertManyRequest returns a new empty InsertManyRequest.
-func MakeInsertManyRequest(space string) InsertManyRequest {
-	req := InsertManyRequest{}
-	req.impl = newCall("crud.insert_many")
-	req.space = space
-	req.opts = InsertManyOpts{}
-	return req
+// NewInsertManyRequest returns a new empty InsertManyRequest.
+func NewInsertManyRequest(space string) InsertManyRequest {
+	return InsertManyRequest{
+		baseRequest: newBaseRequest("crud.insert_many"),
+		space:       space,
+		opts:        InsertManyOpts{},
+	}
 }
 
 // Tuples sets the tuples for the InsertManyRequest request.
@@ -73,8 +74,9 @@ type InsertObjectManyOpts = OperationObjectManyOpts
 // InsertObjectManyRequest helps you to create request object to call
 // `crud.insert_object_many` for execution by a Connection.
 type InsertObjectManyRequest struct {
-	spaceRequest
+	baseRequest
 
+	space   string
 	objects Objects
 	opts    InsertObjectManyOpts
 }
@@ -86,13 +88,13 @@ type insertObjectManyArgs struct {
 	Opts     InsertObjectManyOpts
 }
 
-// MakeInsertObjectManyRequest returns a new empty InsertObjectManyRequest.
-func MakeInsertObjectManyRequest(space string) InsertObjectManyRequest {
-	req := InsertObjectManyRequest{}
-	req.impl = newCall("crud.insert_object_many")
-	req.space = space
-	req.opts = InsertObjectManyOpts{}
-	return req
+// NewInsertObjectManyRequest returns a new empty InsertObjectManyRequest.
+func NewInsertObjectManyRequest(space string) InsertObjectManyRequest {
+	return InsertObjectManyRequest{
+		baseRequest: newBaseRequest("crud.insert_object_many"),
+		space:       space,
+		opts:        InsertObjectManyOpts{},
+	}
 }
 
 // Objects sets the objects for the InsertObjectManyRequest request.