+++ crud/select.go
@@ -85,8 +85,9 @@ func (opts SelectOpts) EncodeMsgpack(enc *msgpack.Encoder) error {
 // SelectRequest helps you to create request object to call `crud.select`
 // for execution by a Connection.
 type SelectRequest struct {
-	spaceRequest
+	baseRequest
 
+	space      string
 	conditions []Condition
 	opts       SelectOpts
 }
@@ -98,14 +99,13 @@ type selectArgs struct {
 	Opts       SelectOpts
 }
 
-// MakeSelectRequest returns a new empty SelectRequest.
-func MakeSelectRequest(space string) SelectRequest {
-	req := SelectRequest{}
-	req.impl = newCall("crud.select")
-	req.space = space
-	req.conditions = nil
-	req.opts = SelectOpts{}
-	return req
+// NewSelectRequest returns a new empty SelectRequest.
+func NewSelectRequest(space string) SelectRequest {
+	return SelectRequest{
+		baseRequest: newBaseRequest("crud.select"),
+		space:       space,
+		opts:        SelectOpts{},
+	}
 }
 
 // Conditions sets the conditions for the SelectRequest request.