+++ decimal/example_test.go
@@ -39,7 +39,7 @@ func Example() {
 
 	spaceNo := uint32(524)
 
-	number, err := MakeDecimalFromString("-22.804")
+	number, err := NewDecimalFromString("-22.804")
 	if err != nil {
 		log.Fatalf("Failed to prepare test decimal: %s", err)
 	}