+++ request.go
@@ -274,33 +274,23 @@ type baseRequest struct {
 }
 
 // Type returns a IPROTO type for the request.
-func (req *baseRequest) Type() iproto.Type {
+func (req baseRequest) Type() iproto.Type {
 	return req.rtype
 }
 
 // Async returns true if the request does not require a response.
-func (req *baseRequest) Async() bool {
+func (req baseRequest) Async() bool {
 	return req.async
 }
 
 // Ctx returns a context of the request.
-func (req *baseRequest) Ctx() context.Context {
+func (req baseRequest) Ctx() context.Context {
 	return req.ctx
 }
 
 // Response creates a response for the baseRequest.
-func (req *baseRequest) Response(header Header, body io.Reader) (Response, error) {
+func (req baseRequest) Response(header Header, body io.Reader) (Response, error) {
 	return DecodeBaseResponse(header, body)
-}
-
-type spaceRequest struct {
-	baseRequest
-
-	space any
-}
-
-func (req *spaceRequest) setSpace(space any) {
-	req.space = space
 }
 
 func EncodeSpace(res SchemaResolver, enc *msgpack.Encoder, space any) error {
@@ -312,16 +302,6 @@ func EncodeSpace(res SchemaResolver, enc *msgpack.Encoder, space any) error {
 		return err
 	}
 	return nil
-}
-
-type spaceIndexRequest struct {
-	spaceRequest
-
-	index any
-}
-
-func (req *spaceIndexRequest) setIndex(index any) {
-	req.index = index
 }
 
 // authRequest implements IPROTO_AUTH request.
@@ -415,14 +395,14 @@ type PingRequest struct {
 }
 
 // NewPingRequest returns a new PingRequest.
-func NewPingRequest() *PingRequest {
-	req := new(PingRequest)
-	req.rtype = iproto.IPROTO_PING
-	return req
+func NewPingRequest() PingRequest {
+	return PingRequest{
+		baseRequest: baseRequest{rtype: iproto.IPROTO_PING},
+	}
 }
 
 // Body fills an msgpack.Encoder with the ping request body.
-func (req *PingRequest) Body(_ SchemaResolver, enc *msgpack.Encoder) error {
+func (req PingRequest) Body(_ SchemaResolver, enc *msgpack.Encoder) error {
 	return enc.EncodeMapLen(0)
 }
 
@@ -432,17 +412,17 @@ func (req *PingRequest) Body(_ SchemaResolver, enc *msgpack.Encoder) error {
 // the timeout option for Connection does not affect the lifetime
 // of the request. For those purposes use context.WithTimeout() as
 // the root context.
-func (req *PingRequest) Context(ctx context.Context) *PingRequest {
+func (req PingRequest) Context(ctx context.Context) PingRequest {
 	req.ctx = ctx
 	return req
 }
 
 // SelectRequest allows you to create a select request object for execution
 // by a Connection.
-
 type SelectRequest struct {
-	spaceIndexRequest
+	baseRequest
 
+	space, index            any
 	isIteratorSet, fetchPos bool
 	offset, limit           uint32
 	iterator                Iter
@@ -450,43 +430,40 @@ type SelectRequest struct {
 }
 
 // NewSelectRequest returns a new empty SelectRequest.
-func NewSelectRequest(space any) *SelectRequest {
-	req := new(SelectRequest)
-	req.rtype = iproto.IPROTO_SELECT
-	req.setSpace(space)
-	req.isIteratorSet = false
-	req.fetchPos = false
-	req.iterator = IterAll
-	req.key = []any{}
-	req.after = nil
-	req.limit = 0xFFFFFFFF
-	return req
+func NewSelectRequest(space any) SelectRequest {
+	return SelectRequest{
+		baseRequest: baseRequest{rtype: iproto.IPROTO_SELECT},
+		space:       space,
+		iterator:    IterAll,
+		key:         []any{},
+		limit:       0xFFFFFFFF,
+	}
 }
 
 // Index sets the index for the select request.
 // Note: default value is 0.
-func (req *SelectRequest) Index(index any) *SelectRequest {
-	req.setIndex(index)
+func (req SelectRequest) Index(index any) SelectRequest {
+	req.index = index
 	return req
 }
 
 // Offset sets the offset for the select request.
 // Note: default value is 0.
-func (req *SelectRequest) Offset(offset uint32) *SelectRequest {
+func (req SelectRequest) Offset(offset uint32) SelectRequest {
 	req.offset = offset
 	return req
 }
 
 // Limit sets the limit for the select request.
 // Note: default value is 0xFFFFFFFF.
-func (req *SelectRequest) Limit(limit uint32) *SelectRequest {
+func (req SelectRequest) Limit(limit uint32) SelectRequest {
 	req.limit = limit
 	return req
 }
 
 // Iterator set the iterator for the select request.
 // Note: default value is IterAll if key is not set or IterEq otherwise.
-func (req *SelectRequest) Iterator(iterator Iter) *SelectRequest {
+func (req SelectRequest) Iterator(iterator Iter) SelectRequest {
 	req.iterator = iterator
 	req.isIteratorSet = true
 	return req
@@ -494,7 +471,7 @@ func (req *SelectRequest) Iterator(iterator Iter) *SelectRequest {
 
 // Key set the key for the select request.
 // Note: default value is empty.
-func (req *SelectRequest) Key(key any) *SelectRequest {
+func (req SelectRequest) Key(key any) SelectRequest {
 	req.key = key
 	if !req.isIteratorSet {
 		req.iterator = IterEq
@@ -510,7 +487,7 @@ func (req *SelectRequest) Key(key any) *SelectRequest {
 // Requires Tarantool >= 2.11.
 //
 // Since 1.11.0.
-func (req *SelectRequest) FetchPos(fetch bool) *SelectRequest {
+func (req SelectRequest) FetchPos(fetch bool) SelectRequest {
 	req.fetchPos = fetch
 	return req
 }
@@ -523,13 +500,13 @@ func (req *SelectRequest) FetchPos(fetch bool) *SelectRequest {
 // Requires Tarantool >= 2.11.
 //
 // Since 1.11.0.
-func (req *SelectRequest) After(after any) *SelectRequest {
+func (req SelectRequest) After(after any) SelectRequest {
 	req.after = after
 	return req
 }
 
 // Body fills an encoder with the select request body.
-func (req *SelectRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
+func (req SelectRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 	spaceEnc, err := newSpaceEncoder(res, req.space)
 	if err != nil {
 		return err
@@ -619,7 +596,7 @@ func (req *SelectRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 // the timeout option for Connection does not affect the lifetime
 // of the request. For those purposes use context.WithTimeout() as
 // the root context.
-func (req *SelectRequest) Context(ctx context.Context) *SelectRequest {
+func (req SelectRequest) Context(ctx context.Context) SelectRequest {
 	req.ctx = ctx
 	return req
 }
@@ -631,7 +608,7 @@ var selectsPool *sync.Pool = &sync.Pool{
 }
 
 // Response creates a response for the SelectRequest.
-func (req *SelectRequest) Response(header Header, body io.Reader) (Response, error) {
+func (req SelectRequest) Response(header Header, body io.Reader) (Response, error) {
 	base, err := createBaseResponse(header, body)
 	if err != nil {
 		return nil, err
@@ -644,33 +621,32 @@ func (req *SelectRequest) Response(header Header, body io.Reader) (Response, err
 }
 
 // InsertRequest helps you to create an insert request object for execution
-
 // by a Connection.
-
 type InsertRequest struct {
-	spaceRequest
+	baseRequest
 
+	space any
 	tuple any
 }
 
 // NewInsertRequest returns a new empty InsertRequest.
-func NewInsertRequest(space any) *InsertRequest {
-	req := new(InsertRequest)
-	req.rtype = iproto.IPROTO_INSERT
-	req.setSpace(space)
-	req.tuple = []any{}
-	return req
+func NewInsertRequest(space any) InsertRequest {
+	return InsertRequest{
+		baseRequest: baseRequest{rtype: iproto.IPROTO_INSERT},
+		space:       space,
+		tuple:       []any{},
+	}
 }
 
 // Tuple sets the tuple for insertion the insert request.
 // Note: default value is nil.
-func (req *InsertRequest) Tuple(tuple any) *InsertRequest {
+func (req InsertRequest) Tuple(tuple any) InsertRequest {
 	req.tuple = tuple
 	return req
 }
 
 // Body fills an msgpack.Encoder with the insert request body.
-func (req *InsertRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
+func (req InsertRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 	spaceEnc, err := newSpaceEncoder(res, req.space)
 	if err != nil {
 		return err
@@ -697,7 +673,7 @@ func (req *InsertRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 // the timeout option for Connection does not affect the lifetime
 // of the request. For those purposes use context.WithTimeout() as
 // the root context.
-func (req *InsertRequest) Context(ctx context.Context) *InsertRequest {
+func (req InsertRequest) Context(ctx context.Context) InsertRequest {
 	req.ctx = ctx
 	return req
 }
@@ -705,29 +681,30 @@ func (req *InsertRequest) Context(ctx context.Context) *InsertRequest {
 // ReplaceRequest helps you to create a replace request object for execution
 // by a Connection.
 type ReplaceRequest struct {
-	spaceRequest
+	baseRequest
 
+	space any
 	tuple any
 }
 
 // NewReplaceRequest returns a new empty ReplaceRequest.
-func NewReplaceRequest(space any) *ReplaceRequest {
-	req := new(ReplaceRequest)
-	req.rtype = iproto.IPROTO_REPLACE
-	req.setSpace(space)
-	req.tuple = []any{}
-	return req
+func NewReplaceRequest(space any) ReplaceRequest {
+	return ReplaceRequest{
+		baseRequest: baseRequest{rtype: iproto.IPROTO_REPLACE},
+		space:       space,
+		tuple:       []any{},
+	}
 }
 
 // Tuple sets the tuple for replace by the replace request.
 // Note: default value is nil.
-func (req *ReplaceRequest) Tuple(tuple any) *ReplaceRequest {
+func (req ReplaceRequest) Tuple(tuple any) ReplaceRequest {
 	req.tuple = tuple
 	return req
 }
 
 // Body fills an msgpack.Encoder with the replace request body.
-func (req *ReplaceRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
+func (req ReplaceRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 	spaceEnc, err := newSpaceEncoder(res, req.space)
 	if err != nil {
 		return err
@@ -754,7 +731,7 @@ func (req *ReplaceRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error 
 // the timeout option for Connection does not affect the lifetime
 // of the request. For those purposes use context.WithTimeout() as
 // the root context.
-func (req *ReplaceRequest) Context(ctx context.Context) *ReplaceRequest {
+func (req ReplaceRequest) Context(ctx context.Context) ReplaceRequest {
 	req.ctx = ctx
 	return req
 }
@@ -762,36 +739,37 @@ func (req *ReplaceRequest) Context(ctx context.Context) *ReplaceRequest {
 // DeleteRequest helps you to create a delete request object for execution
 // by a Connection.
 type DeleteRequest struct {
-	spaceIndexRequest
+	baseRequest
 
-	key any
+	space, index any
+	key          any
 }
 
 // NewDeleteRequest returns a new empty DeleteRequest.
-func NewDeleteRequest(space any) *DeleteRequest {
-	req := new(DeleteRequest)
-	req.rtype = iproto.IPROTO_DELETE
-	req.setSpace(space)
-	req.key = []any{}
-	return req
+func NewDeleteRequest(space any) DeleteRequest {
+	return DeleteRequest{
+		baseRequest: baseRequest{rtype: iproto.IPROTO_DELETE},
+		space:       space,
+		key:         []any{},
+	}
 }
 
 // Index sets the index for the delete request.
 // Note: default value is 0.
-func (req *DeleteRequest) Index(index any) *DeleteRequest {
-	req.setIndex(index)
+func (req DeleteRequest) Index(index any) DeleteRequest {
+	req.index = index
 	return req
 }
 
 // Key sets the key of tuple for the delete request.
 // Note: default value is empty.
-func (req *DeleteRequest) Key(key any) *DeleteRequest {
+func (req DeleteRequest) Key(key any) DeleteRequest {
 	req.key = key
 	return req
 }
 
 // Body fills an msgpack.Encoder with the delete request body.
-func (req *DeleteRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
+func (req DeleteRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 	spaceEnc, err := newSpaceEncoder(res, req.space)
 	if err != nil {
 		return err
@@ -815,7 +793,7 @@ func (req *DeleteRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 // the timeout option for Connection does not affect the lifetime
 // of the request. For those purposes use context.WithTimeout() as
 // the root context.
-func (req *DeleteRequest) Context(ctx context.Context) *DeleteRequest {
+func (req DeleteRequest) Context(ctx context.Context) DeleteRequest {
 	req.ctx = ctx
 	return req
 }
@@ -823,44 +801,45 @@ func (req *DeleteRequest) Context(ctx context.Context) *DeleteRequest {
 // UpdateRequest helps you to create an update request object for execution
 // by a Connection.
 type UpdateRequest struct {
-	spaceIndexRequest
+	baseRequest
 
-	key any
-	ops *Operations
+	space, index any
+	key          any
+	ops          *Operations
 }
 
 // NewUpdateRequest returns a new empty UpdateRequest.
-func NewUpdateRequest(space any) *UpdateRequest {
-	req := new(UpdateRequest)
-	req.rtype = iproto.IPROTO_UPDATE
-	req.setSpace(space)
-	req.key = []any{}
-	return req
+func NewUpdateRequest(space any) UpdateRequest {
+	return UpdateRequest{
+		baseRequest: baseRequest{rtype: iproto.IPROTO_UPDATE},
+		space:       space,
+		key:         []any{},
+	}
 }
 
 // Index sets the index for the update request.
 // Note: default value is 0.
-func (req *UpdateRequest) Index(index any) *UpdateRequest {
-	req.setIndex(index)
+func (req UpdateRequest) Index(index any) UpdateRequest {
+	req.index = index
 	return req
 }
 
 // Key sets the key of tuple for the update request.
 // Note: default value is empty.
-func (req *UpdateRequest) Key(key any) *UpdateRequest {
+func (req UpdateRequest) Key(key any) UpdateRequest {
 	req.key = key
 	return req
 }
 
 // Operations sets operations to be performed on update.
 // Note: default value is empty.
-func (req *UpdateRequest) Operations(ops *Operations) *UpdateRequest {
+func (req UpdateRequest) Operations(ops *Operations) UpdateRequest {
 	req.ops = ops
 	return req
 }
 
 // Body fills an msgpack.Encoder with the update request body.
-func (req *UpdateRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
+func (req UpdateRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 	spaceEnc, err := newSpaceEncoder(res, req.space)
 	if err != nil {
 		return err
@@ -896,7 +875,7 @@ func (req *UpdateRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 // the timeout option for Connection does not affect the lifetime
 // of the request. For those purposes use context.WithTimeout() as
 // the root context.
-func (req *UpdateRequest) Context(ctx context.Context) *UpdateRequest {
+func (req UpdateRequest) Context(ctx context.Context) UpdateRequest {
 	req.ctx = ctx
 	return req
 }
@@ -904,37 +883,38 @@ func (req *UpdateRequest) Context(ctx context.Context) *UpdateRequest {
 // UpsertRequest helps you to create an upsert request object for execution
 // by a Connection.
 type UpsertRequest struct {
-	spaceRequest
+	baseRequest
 
+	space any
 	tuple any
 	ops   *Operations
 }
 
 // NewUpsertRequest returns a new empty UpsertRequest.
-func NewUpsertRequest(space any) *UpsertRequest {
-	req := new(UpsertRequest)
-	req.rtype = iproto.IPROTO_UPSERT
-	req.setSpace(space)
-	req.tuple = []any{}
-	return req
+func NewUpsertRequest(space any) UpsertRequest {
+	return UpsertRequest{
+		baseRequest: baseRequest{rtype: iproto.IPROTO_UPSERT},
+		space:       space,
+		tuple:       []any{},
+	}
 }
 
 // Tuple sets the tuple for insertion or update by the upsert request.
 // Note: default value is empty.
-func (req *UpsertRequest) Tuple(tuple any) *UpsertRequest {
+func (req UpsertRequest) Tuple(tuple any) UpsertRequest {
 	req.tuple = tuple
 	return req
 }
 
 // Operations sets operations to be performed on update case by the upsert request.
 // Note: default value is empty.
-func (req *UpsertRequest) Operations(ops *Operations) *UpsertRequest {
+func (req UpsertRequest) Operations(ops *Operations) UpsertRequest {
 	req.ops = ops
 	return req
 }
 
 // Body fills an msgpack.Encoder with the upsert request body.
-func (req *UpsertRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
+func (req UpsertRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 	spaceEnc, err := newSpaceEncoder(res, req.space)
 	if err != nil {
 		return err
@@ -973,7 +953,7 @@ func (req *UpsertRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 // the timeout option for Connection does not affect the lifetime
 // of the request. For those purposes use context.WithTimeout() as
 // the root context.
-func (req *UpsertRequest) Context(ctx context.Context) *UpsertRequest {
+func (req UpsertRequest) Context(ctx context.Context) UpsertRequest {
 	req.ctx = ctx
 	return req
 }
@@ -989,22 +969,22 @@ type CallRequest struct {
 
 // NewCallRequest returns a new empty CallRequest. It uses request code for
 // Tarantool >= 1.7.
-func NewCallRequest(function string) *CallRequest {
-	req := new(CallRequest)
-	req.rtype = iproto.IPROTO_CALL
-	req.function = function
-	return req
+func NewCallRequest(function string) CallRequest {
+	return CallRequest{
+		baseRequest: baseRequest{rtype: iproto.IPROTO_CALL},
+		function:    function,
+	}
 }
 
 // Args sets the args for the call request.
 // Note: default value is empty.
-func (req *CallRequest) Args(args any) *CallRequest {
+func (req CallRequest) Args(args any) CallRequest {
 	req.args = args
 	return req
 }
 
 // Body fills an encoder with the call request body.
-func (req *CallRequest) Body(_ SchemaResolver, enc *msgpack.Encoder) error {
+func (req CallRequest) Body(_ SchemaResolver, enc *msgpack.Encoder) error {
 	if err := enc.EncodeMapLen(2); err != nil {
 		return err
 	}
@@ -1034,7 +1014,7 @@ func (req *CallRequest) Body(_ SchemaResolver, enc *msgpack.Encoder) error {
 // the timeout option for Connection does not affect the lifetime
 // of the request. For those purposes use context.WithTimeout() as
 // the root context.
-func (req *CallRequest) Context(ctx context.Context) *CallRequest {
+func (req CallRequest) Context(ctx context.Context) CallRequest {
 	req.ctx = ctx
 	return req
 }
@@ -1049,23 +1029,23 @@ type EvalRequest struct {
 }
 
 // NewEvalRequest returns a new empty EvalRequest.
-func NewEvalRequest(expr string) *EvalRequest {
-	req := new(EvalRequest)
-	req.rtype = iproto.IPROTO_EVAL
-	req.expr = expr
-	req.args = []any{}
-	return req
+func NewEvalRequest(expr string) EvalRequest {
+	return EvalRequest{
+		baseRequest: baseRequest{rtype: iproto.IPROTO_EVAL},
+		expr:        expr,
+		args:        []any{},
+	}
 }
 
 // Args sets the args for the eval request.
 // Note: default value is empty.
-func (req *EvalRequest) Args(args any) *EvalRequest {
+func (req EvalRequest) Args(args any) EvalRequest {
 	req.args = args
 	return req
 }
 
 // Body fills an msgpack.Encoder with the eval request body.
-func (req *EvalRequest) Body(_ SchemaResolver, enc *msgpack.Encoder) error {
+func (req EvalRequest) Body(_ SchemaResolver, enc *msgpack.Encoder) error {
 	if err := enc.EncodeMapLen(2); err != nil {
 		return err
 	}
@@ -1095,7 +1075,7 @@ func (req *EvalRequest) Body(_ SchemaResolver, enc *msgpack.Encoder) error {
 // the timeout option for Connection does not affect the lifetime
 // of the request. For those purposes use context.WithTimeout() as
 // the root context.
-func (req *EvalRequest) Context(ctx context.Context) *EvalRequest {
+func (req EvalRequest) Context(ctx context.Context) EvalRequest {
 	req.ctx = ctx
 	return req
 }
@@ -1110,23 +1090,23 @@ type ExecuteRequest struct {
 }
 
 // NewExecuteRequest returns a new empty ExecuteRequest.
-func NewExecuteRequest(expr string) *ExecuteRequest {
-	req := new(ExecuteRequest)
-	req.rtype = iproto.IPROTO_EXECUTE
-	req.expr = expr
-	req.args = []any{}
-	return req
+func NewExecuteRequest(expr string) ExecuteRequest {
+	return ExecuteRequest{
+		baseRequest: baseRequest{rtype: iproto.IPROTO_EXECUTE},
+		expr:        expr,
+		args:        []any{},
+	}
 }
 
 // Args sets the args for the execute request.
 // Note: default value is empty.
-func (req *ExecuteRequest) Args(args any) *ExecuteRequest {
+func (req ExecuteRequest) Args(args any) ExecuteRequest {
 	req.args = args
 	return req
 }
 
 // Body fills an msgpack.Encoder with the execute request body.
-func (req *ExecuteRequest) Body(_ SchemaResolver, enc *msgpack.Encoder) error {
+func (req ExecuteRequest) Body(_ SchemaResolver, enc *msgpack.Encoder) error {
 	if err := enc.EncodeMapLen(2); err != nil {
 		return err
 	}
@@ -1152,7 +1132,7 @@ func (req *ExecuteRequest) Body(_ SchemaResolver, enc *msgpack.Encoder) error {
 // the timeout option for Connection does not affect the lifetime
 // of the request. For those purposes use context.WithTimeout() as
 // the root context.
-func (req *ExecuteRequest) Context(ctx context.Context) *ExecuteRequest {
+func (req ExecuteRequest) Context(ctx context.Context) ExecuteRequest {
 	req.ctx = ctx
 	return req
 }
@@ -1164,7 +1144,7 @@ var executesPool = sync.Pool{
 }
 
 // Response creates a response for the ExecuteRequest.
-func (req *ExecuteRequest) Response(header Header, body io.Reader) (Response, error) {
+func (req ExecuteRequest) Response(header Header, body io.Reader) (Response, error) {
 	baseResp, err := createBaseResponse(header, body)
 	if err != nil {
 		return nil, err
@@ -1185,15 +1165,15 @@ type WatchOnceRequest struct {
 }
 
 // NewWatchOnceRequest returns a new watchOnceRequest.
-func NewWatchOnceRequest(key string) *WatchOnceRequest {
-	req := new(WatchOnceRequest)
-	req.rtype = iproto.IPROTO_WATCH_ONCE
-	req.key = key
-	return req
+func NewWatchOnceRequest(key string) WatchOnceRequest {
+	return WatchOnceRequest{
+		baseRequest: baseRequest{rtype: iproto.IPROTO_WATCH_ONCE},
+		key:         key,
+	}
 }
 
 // Body fills an msgpack.Encoder with the watchOnce request body.
-func (req *WatchOnceRequest) Body(_ SchemaResolver, enc *msgpack.Encoder) error {
+func (req WatchOnceRequest) Body(_ SchemaResolver, enc *msgpack.Encoder) error {
 	if err := enc.EncodeMapLen(1); err != nil {
 		return err
 	}
@@ -1206,7 +1186,7 @@ func (req *WatchOnceRequest) Body(_ SchemaResolver, enc *msgpack.Encoder) error 
 }
 
 // Context sets a passed context to the request.
-func (req *WatchOnceRequest) Context(ctx context.Context) *WatchOnceRequest {
+func (req WatchOnceRequest) Context(ctx context.Context) WatchOnceRequest {
 	req.ctx = ctx
 	return req
 }